+++ rsocket/load_balancer/load_balancer_strategy.py
@@ -4,6 +4,10 @@ from rsocket.rsocket import RSocket
 
 
 class LoadBalancerStrategy(metaclass=abc.ABCMeta):
+    """
+    Base class for load balancer strategies.
+    """
+
     @abc.abstractmethod
     def select(self) -> RSocket:
         ...