+++ rsocket/rsocket_client.py
@@ -24,6 +24,7 @@ class RSocketClient(RSocketBase):
     :param transport_provider: Async generator which returns `Transport` to use with this instance.
     :param request_queue_size: Number of frames which can be queued while waiting for a lease.
     :param fragment_size_bytes: Minimum 64, Maximum depends on transport.
+    :param handler_factory: Callable which returns the implemented application logic endpoints. See also :class:`RequestRouter <rsocket.routing.request_router.RequestRouter>`
     """
 
     def __init__(self,