+++ rsocket/rx_support/rx_handler.py
@@ -11,39 +11,59 @@ from rsocket.logger import logger
 from rsocket.payload import Payload
 from rsocket.rx_support.rx_channel import RxChannel
 
+__all__ = ['RxHandler']
+
 
 class RxHandler:
+    """
+    Variant of :class:`RequestHandler <rsocket.request_handler.RequestHandler>` which uses Rx (3.0).
+    Wrap with :meth:`rx_handler_factory <rsocket.rx_support.rx_handler_adapter.rx_handler_factory>` to pass as a request handler
+    """
 
     @abstractmethod
     async def on_setup(self,
                        data_encoding: bytes,
                        metadata_encoding: bytes,
                        payload: Payload):
-        ...
+        """
+        Handle setup request
+        """
 
     @abstractmethod
     async def on_metadata_push(self, metadata: Payload):
-        ...
+        """
+        Handle metadata-push request
+        """
 
     @abstractmethod
     async def request_channel(self, payload: Payload) -> RxChannel:
-        ...
+        """
+        Handle request-channel interaction
+        """
 
     @abstractmethod
     async def request_fire_and_forget(self, payload: Payload):
-        ...
+        """
+        Handle request-fire-and-forget request
+        """
 
     @abstractmethod
     async def request_response(self, payload: Payload) -> Observable:
-        ...
+        """
+        Handle request-response interaction
+        """
 
     @abstractmethod
     async def request_stream(self, payload: Payload) -> Union[Observable, Callable[[Subject], Observable]]:
-        ...
+        """
+        Handle request-stream interaction
+        """
 
     @abstractmethod
     async def on_error(self, error_code: ErrorCode, payload: Payload):
-        ...
+        """
+        Handle errors received from the remote side
+        """
 
     @abstractmethod
     async def on_keepalive_timeout(self,
@@ -53,11 +73,15 @@ class RxHandler:
 
     @abstractmethod
     async def on_connection_error(self, rsocket, exception: Exception):
-        ...
+        """
+        Handle connection error
+        """
 
     @abstractmethod
     async def on_close(self, rsocket, exception: Optional[Exception] = None):
-        ...
+        """
+        Handle connection closed
+        """
 
     # noinspection PyMethodMayBeStatic
     def _parse_composite_metadata(self, metadata: bytes) -> CompositeMetadata: