+++ rsocket/transports/quart_websocket.py
@@ -12,6 +12,9 @@ from rsocket.transports.abstract_messaging import AbstractMessagingTransport
 async def websocket_handler(on_server_create=None, **kwargs):
     """
     Helper method to instantiate an RSocket server using a quart websocket connection.
+
+    :param on_server_create: callback to be called when the server is created
+    :param kwargs: parameters passed to the server
     """
 
     transport = TransportQuartWebsocket()
@@ -25,7 +28,7 @@ async def websocket_handler(on_server_create=None, **kwargs):
 
 class TransportQuartWebsocket(AbstractMessagingTransport):
     """
-    RSocket transport over server side quart websocket.
+    RSocket transport over server side quart websocket. Use the `websocket_handler <rsocket.transports.quart_websocket.websocket_handler>` helper method to instantiate.
     """
 
     async def handle_incoming_ws_messages(self):