+++ tests/rsocket/test_routing.py
@@ -5,12 +5,14 @@ import pytest
 
 from reactivestreams.subscriber import DefaultSubscriber
 from rsocket.awaitable.awaitable_rsocket import AwaitableRSocket
+from rsocket.error_codes import ErrorCode
 from rsocket.extensions.authentication import Authentication, AuthenticationSimple
 from rsocket.extensions.composite_metadata import CompositeMetadata
 from rsocket.extensions.helpers import route, composite, authenticate_simple
 from rsocket.extensions.mimetypes import WellKnownMimeTypes
 from rsocket.helpers import create_response
 from rsocket.payload import Payload
+from rsocket.request_handler import BaseRequestHandler
 from rsocket.routing.request_router import RequestRouter
 from rsocket.routing.routing_request_handler import RoutingRequestHandler
 from rsocket.rx_support.rx_rsocket import RxRSocket
@@ -357,3 +359,33 @@ async def test_valid_authentication_in_routing_handler(lazy_pipe):
                                                                                                          'pass'))))
 
         assert result.data == b'result'
+
+
+@pytest.mark.allow_error_log(regex_filter='(RSocket error REJECTED_SETUP|Setup error)')
+async def test_invalid_metadata_for_routing(lazy_pipe):
+    router = RequestRouter()
+
+    async def authenticate(path: str, authentication: Authentication):
+        if not isinstance(authentication, AuthenticationSimple) or authentication.password != b'pass':
+            raise Exception('Invalid credentials')
+
+    error_wait = asyncio.Event()
+
+    def client_handler_factory():
+        class ClientHandler(BaseRequestHandler):
+            async def on_error(self, error_code: ErrorCode, payload: Payload):
+                error_wait.set()
+
+        return ClientHandler()
+
+    @router.response('test.path')
+    async def response():
+        return create_response(b'result')
+
+    def handler_factory():
+        return RoutingRequestHandler(router, authentication_verifier=authenticate)
+
+    async with lazy_pipe(
+            client_arguments={'handler_factory': client_handler_factory},
+            server_arguments={'handler_factory': handler_factory}) as (server, client):
+        await error_wait.wait()